Store a copy of the exported tree, instead of keeping the existing object intact.

+++ fcrepo-kernel-impl/src/main/java/org/fcrepo/kernel/impl/FedoraBinaryImpl.java
@@ -312,7 +312,9 @@ public class FedoraBinaryImpl extends FedoraResourceImpl implements FedoraBinary
     public void delete() {
         final Datastream description = getDescription();
 
-        super.delete();
+        if (!softDelete()) {
+            super.delete();
+        }
 
         description.delete();
     }
